+++ fitshow/src/fit.rs
@@ -3,97 +3,129 @@ use utilities::FITActivity;
 pub fn print_activity(act: &FITActivity, detailed: bool) {
     let unknown = "Unknown".to_string();
 
-    println!("\n:: FILE: {}", act.session.filename.as_ref().unwrap());
     println!(
-        "Manufacturer:  {}",
+        "\nFile:                     {}",
+        act.session.filename.as_ref().unwrap()
+    );
+    println!(
+        "Manufacturer:             {}",
         act.session.manufacturer.as_ref().unwrap_or(&unknown)
     );
     println!(
-        "Product:       {}",
+        "Product:                  {}",
         act.session.product.as_ref().unwrap_or(&unknown)
     );
     println!(
-        "Serial number: {}",
+        "Serial number:            {}",
         act.session.serial_number.as_ref().unwrap_or(&unknown)
     );
     println!(
-        "Time created:  {}",
+        "Time created:             {}",
         act.session.time_created.as_ref().unwrap()
     );
     println!(
-        "Activity type:   {}",
+        "Activity type:            {}",
         act.session.activity_type.as_ref().unwrap_or(&unknown)
     );
     println!(
-        "Activity detail: {}",
+        "Activity detail:          {}",
         act.session.activity_detailed.as_ref().unwrap_or(&unknown)
     );
     println!(
-        "Sessions:           {:5}",
+        "Sessions:                 {}",
         act.session.num_sessions.unwrap_or_default()
     );
     println!(
-        "Laps:               {:5}",
+        "Laps:                     {}",
         act.session.num_laps.unwrap_or_default()
     );
     println!(
-        "Records:            {:5}",
+        "Records:                  {}",
         act.session.num_records.unwrap_or_default()
     );
-    println!("Total duration:  {}", act.session.duration.unwrap());
     println!(
-        "Calories Burned:    {:5}",
+        "Total duration:           {}",
+        act.session.duration.unwrap()
+    );
+    println!(
+        "Calories Burned:          {}",
         act.session.calories.unwrap_or_default()
     );
 
     println!(
-        "Cadence Avg: {}, Max: {}",
-        act.session.cadence_avg.unwrap_or_default(),
+        "Cadence Avg:              {}",
+        act.session.cadence_avg.unwrap_or_default()
+    );
+    println!(
+        "Cadence Max:              {}",
         act.session.cadence_max.unwrap_or_default()
     );
     println!(
-        "Heart Rate Min: {}, Avg: {}, Max: {}",
-        act.session.heartrate_min.unwrap_or_default(),
-        act.session.heartrate_avg.unwrap_or_default(),
+        "Heart Rate Min:           {}",
+        act.session.heartrate_min.unwrap_or_default()
+    );
+    println!(
+        "Heart Rate Avg:           {}",
+        act.session.heartrate_avg.unwrap_or_default()
+    );
+    println!(
+        "Heart Rate Max:           {}",
         act.session.heartrate_max.unwrap_or_default()
     );
 
     println!(
-        "Speed Avg: {:?} m/s, Max: {:?} m/s",
-        act.session.speed_avg.unwrap_or_default().value,
+        "Speed Avg:                {:?} m/s",
+        act.session.speed_avg.unwrap_or_default().value
+    );
+    println!(
+        "Speed Max:                {:?} m/s",
         act.session.speed_max.unwrap_or_default().value
     );
+
     println!(
-        "Power Avg: {}, Max: {}, Threshold: {}",
-        act.session.power_avg.unwrap_or_default(),
-        act.session.power_max.unwrap_or_default(),
-        act.session.power_threshold.unwrap_or_default()
+        "Power Avg:                {}",
+        act.session.power_avg.unwrap_or_default()
     );
     println!(
-        "Ascent:   {:5} m",
+        "Power Max:                {}",
+        act.session.power_max.unwrap_or_default()
+    );
+    println!(
+        "Power Threshold:          {}",
+        act.session.power_threshold.unwrap_or_default()
+    );
+
+    println!(
+        "Ascent:                   {} m",
         act.session.ascent.unwrap_or_default().value
     );
     println!(
-        "Descent:  {:5} m",
+        "Descent:                  {} m",
         act.session.descent.unwrap_or_default().value
     );
     println!(
-        "Distance: {:5} m",
+        "Distance:                 {} m",
         act.session.distance.unwrap_or_default().value
     );
     if detailed {
         println!(
-            "North East Latitude: {}, Longitude: {}",
-            act.session.nec_lat.unwrap_or_default(),
+            "North East Latitude:      {}",
+            act.session.nec_lat.unwrap_or_default()
+        );
+        println!(
+            "North East Longitude:     {}",
             act.session.nec_lon.unwrap_or_default()
         );
         println!(
-            "South West Latitude: {}, Longitude: {}",
-            act.session.swc_lat.unwrap_or_default(),
+            "South West Latitude:      {}",
+            act.session.swc_lat.unwrap_or_default()
+        );
+        println!(
+            "South West Longitude:     {}",
             act.session.swc_lon.unwrap_or_default()
         );
         println!(
-            "Stance time avg: {}s",
+            "Stance time avg:          {}s",
             act.session.stance_time_avg.unwrap_or_default()
         );
         println!(
@@ -101,34 +133,40 @@ pub fn print_activity(act: &FITActivity, detailed: bool) {
             act.session.vertical_oscillation_avg.unwrap_or_default()
         );
         println!(
-            "Duration Active: {}",
+            "Duration Active:          {}",
             act.session.duration_active.unwrap_or_default()
         );
         println!(
-            "Duration Moving: {}",
+            "Duration Moving:          {}",
             act.session.duration_moving.unwrap_or_default()
         );
-        println!("Start time:  {}", act.session.start_time.as_ref().unwrap());
-        println!("Finish time: {}", act.session.finish_time.as_ref().unwrap());
+        println!(
+            "Start time:               {}",
+            act.session.start_time.as_ref().unwrap()
+        );
+        println!(
+            "Finish time:              {}",
+            act.session.finish_time.as_ref().unwrap()
+        );
         println!("Time in Zones:");
         println!(
-            "  Speed/Power: {}",
+            "  Speed/Power:            {}",
             act.session.time_in_hr_zones.hr_zone_4.unwrap()
         );
         println!(
-            "  Anaerobic:   {}",
+            "  Anaerobic:              {}",
             act.session.time_in_hr_zones.hr_zone_3.unwrap()
         );
         println!(
-            "  Aerobic:     {}",
+            "  Aerobic:                {}",
             act.session.time_in_hr_zones.hr_zone_2.unwrap()
         );
         println!(
-            "  Fat Burning: {}",
+            "  Fat Burning:            {}",
             act.session.time_in_hr_zones.hr_zone_1.unwrap()
         );
         println!(
-            "  Warmup:      {}",
+            "  Warmup:                 {}",
             act.session.time_in_hr_zones.hr_zone_0.unwrap()
         );
     }